requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv==0.15.0
pytest
gunicorn
flasgger
quart-cors
Flask-Migrate
Flask-Login
gino==1.0.1
quart==0.15.1
quart-auth==0.6.0
gino-quart==0.1.2
asyncpg>=0.25.0
aiofiles==0.6.0
blinker==1.4
click==8.0.0
hypercorn==0.11.2
itsdangerous==1.1.0
jinja2==2.11.3
toml==0.10.2
werkzeug>=2.0.0
h11==0.9.0
h2==3.2.0
priority==1.3.0
wsproto==0.15.0
SQLAlchemy==1.3.24
alembic==1.4.3
markupsafe==2.0.1
importlib_metadata==3.10.1
setuptools>=57.0.0
